% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vam.R
\name{vam}
\alias{vam}
\title{Variational Additive Model}
\usage{
vam(
  fixed,
  se = NULL,
  fe = NULL,
  Ts = NULL,
  data,
  se.spline = NULL,
  fe.spline = NULL,
  family = stats::gaussian(link = "identity"),
  priors = list(Ae = 0.01, Be = 0.01, Af = c(0.01, 0.01), Bf = c(0.01, 0.01), sigB =
    1000),
  tol = 1e-05,
  maxIter = 500,
  up = 100,
  dots = up/10,
  verbose = TRUE
)
}
\arguments{
\item{fixed}{an object of class \link{formula} containing the fixed effects.}

\item{se}{a one-sided object of class \link{formula} containing the smoothed effects.}

\item{fe}{a one-sided object of class \link{formula} containing the functional effects.}

\item{Ts}{a vector containing the grid on which the function is measured.}

\item{data}{\link{data.frame} containing the variables in the fixed effects model.}

\item{se.spline}{a list containing three objects specifying the spline for the smoothed effects: \code{df} degrees of freedom,
\code{degree} of the spline, and \code{slength} the number of evaluation points to use when graphing the smoothed effect.
Defaults to \code{list(df = 8, degree = 3, slength = 100)}.}

\item{fe.spline}{a \link{list} containing three objects specifying the spline for the functional effect:  \code{df} degrees of freedom,
\code{degree} of the spline, and \code{xi} a sparsity control between 0 and 1. Defaults to \code{list(df = 16, degree = 3, xi = 0.01)}.}

\item{family}{a description of the error distribution and link function to be used in the model.
Currently only \code{gaussian(link = "identity")} and \code{binomial(link = "probit")} are supported.}

\item{priors}{a \link{list} containing the prior specifications for the model errors (\code{Ae} and \code{Be}),
the functional or smoothed effects (\code{Af} or \code{Bf}, can be vectors to vary priors on smoothed tuning parameters),
and the variance of the fixed effects (\code{sigB}). Defaults to
\code{list(Ae = 0.01, Be = 0.01, Af = c(0.01, 0.01), Bf = c(0.01, 0.01), sigB = 1000)}.}

\item{tol}{\link{numeric} value indicating the threshold or tolerance at which the model is judged to have converged. Defaults to 1e-5.}

\item{maxIter}{\link{numeric} value indicating the maximum number of iterations to perform should \code{tol} not be reached. Defaults to 500.}

\item{up}{an \link{integer} value dictating the frequency of printed iteration update statements. Defaults to 100.}

\item{dots}{an \link{integer} value dictating the frequency of printed iteration update dots. Defaults to 10.}

\item{verbose}{a \link{logical} value. If \code{TRUE}, iteration updates are printed in the console. If \code{FALSE}, they are suppressed.
Defaults to \code{TRUE}.}
}
\value{
A list with class "\code{vam}" containing the following components:\tabular{ll}{
\code{VBEstimates} \tab The last iteration of variational Bayesian estimates of the mean model. \cr
\tab \cr
\code{Muqa} \tab  if family is binomial, the last latent vector otherise \link{NULL}. \cr
\tab \cr
\code{fixedEffects} \tab estimates of \eqn{\boldsymbol{\beta}}. \cr
\tab \cr
\code{smoothEffect} \tab estimates of the smoothed effects. \cr
\tab \cr
\code{funcEffect} \tab estimates of the functional effects. \cr
\tab \cr
\code{Ctf} \tab last vector \eqn{c(t)} for use in ZLS test. \cr
\tab \cr
\code{data.name} \tab a character string giving the names of the data. \cr
\tab \cr
\code{varComps} \tab a list containing model variance estimates: \code{Sigmaq} last covariance,
\code{Bqe} last component of model variance, and \code{Bqf} last components of smoothed/functional effects. \cr
\tab \cr
\code{model} \tab a list containing model details including: arguments \code{fixed}, \code{fe}, \code{se}, \code{data}, and \code{family}
as well as the number of fixed effects \code{p}, number of spline bases \code{Kf} (total) and \code{Kfi} (by effect), the design matrix (\code{dmat}),
penalty matrices (\code{O}), basis matrices (\code{Th}), the outcome \code{Y}, and a vector indicating effect type for testing (\code{test.flag}). \cr
\tab \cr
\code{priors} \tab argument \code{prior}. \cr
\tab \cr
\code{nfun} \tab total number of functions fit, smoothed and functional. \cr
\tab \cr
\code{algOut} \tab list of algorithm details including: total iterations to convergence (\code{totalIter}), \code{tol}, \code{maxIter}, and
a list (\code{logpd}) containing details on the changes in the K-L lower bound
(\code{lDelta} tracks changes, \code{logp} tracks current lower bound, \code{prev_logp} tracks previous). \cr
\tab \cr
\code{se.spline} \tab argument \code{se.spline}. \cr
\tab \cr
\code{fe.spline} \tab argument \code{fe.spline}. \cr
}
}
\description{
Implements variational algorithms for estimating effects from both Gaussian and binary additive models with
arbitrary numbers of smoothed/semiparametric and functional effects.
}
\details{
Varitaional additive models are additive models fit using variational Bayesian inference. The models generally have the form
\deqn{g(y_i) = \alpha + \boldsymbol{\beta}'\textbf{x}_i + \sum_{m = 1}^M s_m(z_{im}) + \sum_{f = 1}^F \int_{t\in\mathcal{T}} w_{if}(t)\gamma_f(t) dt + \epsilon_i,}
where \eqn{s()} and \eqn{\gamma()} are smoothed and functional effects, respectively. \eqn{g()} is a link function determined by the
arugment \code{family}. Currently, \code{vam} only supports \code{family = gaussian(link = "identity")} or \code{family = binomial(link = "probit")}.
Variational inference is a deterministic, algorithmic approach to estimating Bayesian posterior densities.
Both smoothed and functional effects are penalized via the mixed model representation of semiparametric regression effects.

Each model component is specified separately using the \code{fixed} argument for fixed effects (i.e. non-penalized effects), \code{se} for
smoothed effects, and \code{fe} for functional effects.
}
\examples{
##### Semiparametric Effect #####
library(SemiPar)
data(lidar)

modLi   <- vam(fixed = logratio ~ 1, se = ~ range, data = lidar,
               se.spline = list(df = 8, degree = 3, slength = 100))
plot(modLi)
coef(modLi)

##### Functional Effect #####
library(refund)
DTI2_1v         <- DTI2[which(DTI2$visit == 1), ]
DTI2_1v$pasatb  <- 1*(DTI2_1v$pasat > median(DTI2_1v$pasat))
cca             <- refund::fpca.sc(DTI2_1v$cca)
fcca            <- cca$Yhat

###### CCA ######
modCc   <- vam(fixed = pasatb ~ 1, fe = ~ fcca, data = DTI2_1v, Ts = list(fcca = ncol(fcca)),
               family = binomial(link = "probit"),
               fe.spline = list(df = 9, degree = 3, xi = 0.01))
plot(modCc)
coef(modCc)

}
\references{
Meyer, MJ and Wei, J (2024). Global tests for smoothed functions in mean field variational additive models. \emph{Available on arXiv} at \url{https://arxiv.org/abs/2406.08168}
}
